refactoring - redemption does not work
