+++ src/char/hammerdin.py
@@ -1,3 +1,4 @@
+from logging import Filter
 import keyboard
 from ui import skills
 from utils.custom_mouse import mouse
@@ -5,7 +6,7 @@ from char import IChar, CharacterCapabilities
 from template_finder import TemplateFinder
 from pather import Pather
 from logger import Logger
-from screen import convert_abs_to_monitor, grab, convert_screen_to_abs
+from screen import convert_abs_to_monitor
 from config import Config
 from utils.misc import wait
 import time
@@ -13,6 +14,7 @@ from pather import Pather, Location
 import cv2 #for Diablo
 from item.pickit import PickIt #for Diablo
 import numpy as np
+from mob_detection import mobcheck
 
 
 class Hammerdin(IChar):
@@ -172,34 +174,6 @@ class Hammerdin(IChar):
      # Chaos Sanctuary, Trash, Seal Bosses (a = Vizier, b = De Seis, c = Infector) & Diablo #
      ########################################################################################
 
-    def mobcheck(self) -> bool:
-        #wait(1) # let the merc paint some mobs
-        self.image = grab()
-        input = self.image
-        cv2.imwrite(f"./info_screenshots/info_dia_mob_" + time.strftime("%Y%m%d_%H%M%S") + ".png", self.image)
-        filterimage, threshz = TemplateFinder().apply_filter(self.image, mask_char=True, mask_hud=True, info_ss=False, erode=0, dilate=2, blur=4, lh=35, ls=0, lv=43, uh=133, us=216, uv=255, bright=255, contrast=139, thresh=10, invert=0) # add HSV filter for #holy freeze & conviction or poison
-        pos_marker = []
-        pos_rectangle = []
-        filterimage, pos_rectangle, pos_marker = TemplateFinder().add_markers(filterimage, threshz, info_ss=False, rect_min_size=100, rect_max_size=200, marker=True)
-        cv2.imwrite(f"./info_screenshots/info_dia_mob__filtered" + time.strftime("%Y%m%d_%H%M%S") + ".png", filterimage)
-        order = TemplateFinder().get_targets_ordered_by_distance(pos_marker, 150)
-        if not order:
-            Logger.info('\033[93m' + "Mobdetection: no Mob detected" + '\033[0m')
-            return False
-        else:
-            pos_m = convert_screen_to_abs(order[0]) #nearest marker
-            pos_m = convert_abs_to_monitor(pos_m)
-            Logger.debug('\033[92m' + "Mobdetection: Found Mob at " + str(pos_m) + " attacking now!" + '\033[0m')
-            pt2 = (640,360)
-            x1, y1 = order[0]
-            pt1 = (int(x1),int(y1))
-            input = np.ascontiguousarray(input)
-            cv2.arrowedLine(input, pt2, pt1, line_type=cv2.LINE_4, thickness=2, tipLength=0.3, color=(255, 0, 255))
-            cv2.imwrite(f"./info_screenshots/info_dia_mob_add_line" + time.strftime("%Y%m%d_%H%M%S") + ".png", input)
-            #cv2.imwrite(f"./info_screenshots/info_cowfound_" + time.strftime("%Y%m%d_%H%M%S") + ".png", self.image)
-            #cv2.imwrite(f"./info_screenshots/info_cowfound_marker_" + time.strftime("%Y%m%d_%H%M%S") + ".png", filterimage)
-            return True
-
     def kill_cs_trash(self, location:str) -> bool:
 
         ###########
@@ -209,7 +183,7 @@ class Hammerdin(IChar):
         if location == "sealdance": #if seal opening fails & trash needs to be cleared -> used at ANY seal
             ### APPROACH
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -234,7 +208,7 @@ class Hammerdin(IChar):
             pos_m = convert_abs_to_monitor((0, 0))
             ### ATTACK ###
             wait(1)#give merc the chance to activate holy freeze
-            if self.mobcheck():
+            if mobcheck(self):
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
                 self._cast_hammers(0.75, "redemption")
@@ -256,7 +230,7 @@ class Hammerdin(IChar):
             if not self._pather.traverse_nodes([604], self, time_out=3): return False  #threshold=0.8 (ex 601)
             ### ATTACK ###
             wait(1)#give merc the chance to activate holy freeze
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -276,7 +250,7 @@ class Hammerdin(IChar):
             self._pather.traverse_nodes_fixed("diablo_entrance_hall_1", self) # 604 -> 671 Hall1
             ### ATTACK ###
             wait(1)#give merc the chance to activate holy freeze
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -297,7 +271,7 @@ class Hammerdin(IChar):
             self._pather.traverse_nodes_fixed("diablo_entrance_1_670_672", self) # 604 -> 671 Hall1
             if not self._pather.traverse_nodes([670], self): return False # pull top mobs 672 to bottom 670
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -322,7 +296,7 @@ class Hammerdin(IChar):
         elif location == "entrance1_01": #static_path "diablo_entrance_hall_2", Hall1 (before layout check)
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -341,7 +315,7 @@ class Hammerdin(IChar):
         elif location == "entrance1_02": #node 673
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -361,7 +335,7 @@ class Hammerdin(IChar):
         elif location == "entrance1_03": #node 674
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -383,7 +357,7 @@ class Hammerdin(IChar):
         elif location == "entrance1_04": #node 676- Hall3
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -404,7 +378,7 @@ class Hammerdin(IChar):
         elif location == "entrance2_01": #static_path "diablo_entrance_hall_2"
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -426,7 +400,7 @@ class Hammerdin(IChar):
             wait (0.2, 0.5)
             if not self._pather.traverse_nodes([605], self): return False#, time_out=3)
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -451,7 +425,7 @@ class Hammerdin(IChar):
             self._pather.traverse_nodes_fixed("diablo_trash_b_hall2_605_top2", self) #pull mobs from top
             if not self._pather.traverse_nodes([605], self): return False#, time_out=3)
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -477,7 +451,7 @@ class Hammerdin(IChar):
             self._pather.traverse_nodes_fixed("diablo_trash_b_hall3_pull_609", self)
             if not self._pather.traverse_nodes([609], self): return False#, time_out=3)
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((0, 0), Config().char["atk_len_cs_trashmobs"])
@@ -506,7 +480,7 @@ class Hammerdin(IChar):
         elif location == "dia_trash_a": #trash before between Pentagramm and Seal A Layoutcheck
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -525,7 +499,7 @@ class Hammerdin(IChar):
         elif location == "dia_trash_b": #trash before between Pentagramm and Seal B Layoutcheck
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -544,7 +518,7 @@ class Hammerdin(IChar):
         elif location == "dia_trash_c": ##trash before between Pentagramm and Seal C Layoutcheck
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"])
@@ -567,7 +541,7 @@ class Hammerdin(IChar):
         elif location == "layoutcheck_a": #layout check seal A, node 619 A1-L, node 620 A2-Y
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -586,7 +560,7 @@ class Hammerdin(IChar):
         elif location == "layoutcheck_b": #layout check seal B, node 634 B1-S, node 649 B2-U
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -604,7 +578,7 @@ class Hammerdin(IChar):
         elif location == "layoutcheck_c": #layout check seal C, node 656 C1-F, node 664 C2-G
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -632,7 +606,7 @@ class Hammerdin(IChar):
         elif location == "pent_before_b": #node 602, pentagram, before CTA buff & depature to layout check
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -650,7 +624,7 @@ class Hammerdin(IChar):
         elif location == "pent_before_c": #node 602, pentagram, before CTA buff & depature to layout check
             ### APPROACH ###
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -674,7 +648,7 @@ class Hammerdin(IChar):
             if not self._pather.traverse_nodes([611], self): return False # , time_out=3):
             ### ATTACK ###
             wait(1)#give merc the chance to activate holy freeze
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -693,7 +667,7 @@ class Hammerdin(IChar):
             ### APPROACH ###
             if not self._pather.traverse_nodes([612], self): return False # , time_out=3):
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 if self._skill_hotkeys["redemption"]:
                     keyboard.send(self._skill_hotkeys["redemption"])
                     wait(0.3, 0.6)
@@ -717,7 +691,7 @@ class Hammerdin(IChar):
             ### APPROACH ###
             if not self._pather.traverse_nodes([613], self): return False # , time_out=3):
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -766,7 +740,7 @@ class Hammerdin(IChar):
             #if not self._pather.traverse_nodes([622], self): return False # , time_out=3):
             if not self._pather.traverse_nodes([622], self): return False
             wait(1)#give merc the chance to activate holy freeze
-            if self.mobcheck():
+            if mobcheck(self):
                 if self._skill_hotkeys["redemption"]:
                     keyboard.send(self._skill_hotkeys["redemption"])
                     wait(0.3, 0.6)
@@ -789,7 +763,7 @@ class Hammerdin(IChar):
             ### APPROACH ###
             # if not self._pather.traverse_nodes([623,624], self): return False #
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -935,7 +909,7 @@ class Hammerdin(IChar):
             wait(0.1,0.3)
             if not self._pather.traverse_nodes([655], self): return False # , time_out=3):
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -959,7 +933,7 @@ class Hammerdin(IChar):
             self._pather.traverse_nodes_fixed("dia_c1f_654_651", self)
             if not self._pather.traverse_nodes([652], self): return False # , time_out=3):
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -1036,7 +1010,7 @@ class Hammerdin(IChar):
             seal_layout="C2-G"
             self._pather.traverse_nodes_fixed("dia_c2g_663", self)
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 Logger.debug(seal_layout + ": Attacking Infector at position 1/1")
@@ -1058,7 +1032,7 @@ class Hammerdin(IChar):
             ### APPROACH ###
             Logger.debug("I have no location argument given for kill_cs_trash(" + location + "), should not happen. Throwing some random hammers")
             ### ATTACK ###
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_cs_trashmobs"] * 0.5)
@@ -1082,7 +1056,7 @@ class Hammerdin(IChar):
             if not self._pather.traverse_nodes([612], self): return False # , time_out=3):
             ### ATTACK ###
             Logger.debug(seal_layout + ": Attacking Vizier at position 1/2")
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_vizier"] * 0.5)
@@ -1090,7 +1064,7 @@ class Hammerdin(IChar):
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking Vizier at position 2/2")
             self._pather.traverse_nodes([611], self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_vizier"] * 0.5)
                 self._move_and_attack((-30, -15), Config().char["atk_len_diablo_vizier"]) # no factor, so merc is not reset by teleport and he his some time to move & kill stray bosses
                 self._cast_hammers(1, "redemption")
@@ -1115,7 +1089,7 @@ class Hammerdin(IChar):
             if not self._pather.traverse_nodes([627, 622], self): return False # , time_out=3):
             ### ATTACK ###
             Logger.debug(seal_layout + ": Attacking Vizier at position 1/2")
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_vizier"] * 0.5)
@@ -1123,13 +1097,13 @@ class Hammerdin(IChar):
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking Vizier at position 2/2")
             self._pather.traverse_nodes([623], self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_vizier"] * 0.5)
                 self._move_and_attack((-30, -15), Config().char["atk_len_diablo_vizier"] * 0.5)
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking Vizier at position 3/3")
             if not self._pather.traverse_nodes([624], self): return False
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_vizier"] * 0.5)
                 self._move_and_attack((-30, -15), Config().char["atk_len_diablo_vizier"])
                 wait(0.1, 0.15)
@@ -1169,7 +1143,7 @@ class Hammerdin(IChar):
             nodes3 = [632]
             ### ATTACK ###
             Logger.debug(seal_layout + ": Attacking De Seis at position 1/4")
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_deseis"] * 0.2)
@@ -1177,18 +1151,18 @@ class Hammerdin(IChar):
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking De Seis at position 2/4")
             self._pather.traverse_nodes(nodes1, self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_deseis"] * 0.2)
                 self._move_and_attack((-30, -15), Config().char["atk_len_diablo_deseis"] * 0.2)
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking De Seis at position 3/4")
             self._pather.traverse_nodes(nodes2, self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((0, 0), Config().char["atk_len_diablo_deseis"] * 0.5)
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking De Seis at position 4/4")
             self._pather.traverse_nodes(nodes3, self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((0, 0), Config().char["atk_len_diablo_deseis"])  # no factor, so merc is not reset by teleport and he his some time to move & kill stray bosses
                 wait(0.1, 0.2)
                 self._cast_hammers(2, "redemption")
@@ -1210,7 +1184,7 @@ class Hammerdin(IChar):
             nodes3 = [641]
             ### ATTACK ###
             Logger.debug(seal_layout + ": Attacking De Seis at position 1/4")
-            if self.mobcheck():
+            if mobcheck(self):
                 pos_m = convert_abs_to_monitor((0, 0))
                 mouse.move(*pos_m, randomize=80, delay_factor=[0.5, 0.7])
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_deseis"] * 0.2)
@@ -1218,18 +1192,18 @@ class Hammerdin(IChar):
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking De Seis at position 2/4")
             self._pather.traverse_nodes(nodes1, self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((30, 15), Config().char["atk_len_diablo_deseis"] * 0.2)
                 self._move_and_attack((-30, -15), Config().char["atk_len_diablo_deseis"] * 0.2)
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking De Seis at position 3/4")
             self._pather.traverse_nodes(nodes2, self, time_out=3)
-            if self.mobcheck():
+            if mobcheck(self):
                 self._move_and_attack((0, 0), Config().char["atk_len_diablo_deseis"] * 0.5)
                 self._cast_hammers(1, "redemption")
             Logger.debug(seal_layout + ": Attacking De Seis at position 4/4")
             self._pather.traverse_nodes(nodes3, self, time_out=3)
-            if self.mobcheck():                
+            if mobcheck(self):                
                 self._move_and_attack((0, 0), Config().char["atk_len_diablo_deseis"])  # no factor, so merc is not reset by teleport and he his some time to move & kill stray bosses
                 wait(0.1, 0.2)
                 self._cast_hammers(2, "redemption")
